Refactoring (Presentation Layer folder)
