+++ lib/presentation/welcome/welcome_screen.dart
@@ -1,4 +1,4 @@
-import 'package:arithmetic_pvp/authentication/login.dart';
+import 'package:arithmetic_pvp/presentation/authentication/login.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
